LaunchMyRoutine.bat
Extensive and Comprehensive List of examples you can put into a batch file, you can use batch files to reduce the amount of opening and searching for folders. It was driving me nuts not being able to find some folders. Be careful not to execute too many files in a single batch. Rather set up a batch file for each workflow you have where the amount of clicking is A LOT.